refactor(service): created service folder and moved user service in service folder
