+++ src/app/app.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { UserService } from 'src/Services/user.service';
+
 
 @Component({
   selector: 'app-root',
