+++ src/app/login-page/login-page/login-page.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { Router, Routes } from '@angular/router';
-import { UserService } from 'src/Services/user.service';
+
 import { RegisterPageComponent } from 'src/app/register-page/register-page.component';
 
 @Component({
@@ -9,7 +9,7 @@ import { RegisterPageComponent } from 'src/app/register-page/register-page.compo
   styleUrls: ['./login-page.component.scss']
 })
 export class LoginPageComponent {
-  constructor(private router: Router, private UserService: UserService) { }
+  constructor(private router: Router, ) { }
 
   navigateToRegister(): void {
     this.router.navigate(['/register']);
